curve: PN triangle: try to calculate normal using bary coords
This was a pretty good solution but it cannot work. See below for
details.

Bary coords would have worked if the points generated were in the same
plane as the original triangle, since they are not, the bary coords
aren't correct and don't follow the curvature variation. For correct
results, need to calculate these by hand.

TODO(ish): correctly calculate the normals of the new points generated.