curve: PN triangles: generate control points not more PN triangles
The way PN triangles work is that they help generate the cubic bezier
triangle's control points using just the vertex positions and its
normals. This information is then used to actually subdivide the
triangle to generate the curved surface that is needed.

The control points generated from the PN triangle must not be thought
of as the new points to further subdivide that but must only be used
as the control points in a cubic bezier surface path.