main: temporary fix to make mesh index slider better to use
In case there is an error while visualizing the mesh, the error
causes the UI elements downwards which makes scrubbing through the
meshes harder when an error shows up. This helps reduce that but still
has the problem of the elements moving up and down.

The true fix would be to have the errors show up on the side and get
recorded into a load that is then displayed on queue.