adaptive mesh: visualize flippable edges with anisotropic criterion
This is a very hacky way to do it. I really hate it but right now
because of time constraint, forced to do it this way where the traits
are all not defined properly and this will lead to issues later on
when adding more features.

Will definitely need to rewrite a bunch of this properly.