config: mouse edge select: use ray direction instead of camera front
The camera front is best only if the section is going to be in the
middle of the screen.

For getting the maximum surface area (which is what a cylinder would
do) we need to take the ray direction when calculating the edge
normal.

It is weird to think about it but the visualization of the edge's
triangles generated for selecting the edge really helps.