mesh: visualize config: sew edge related: must have connecting edge
For an edge to be an edge between sewing edges, the other vertices of
the sewing edges should have a connecting edge. So it should form a
loop consisting of the edge in question, then a loose edge, then the
connecting edge and finally a loose edge which connects back to the
edge in question.

	 edge
       _________
      |	     	|
loose |		|loose
edge 1|	       	|edge 2
      |_________|
     connecting edge